backend_ml/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
starlette
